+++ StereoKitC/intersect.cpp
@@ -46,7 +46,7 @@ bool32_t sphere_ray_intersect(sphere_t sphere, ray_t ray, vec3 *out_pt) {
 	float c = vec3_dot(oc, oc) - sphere.radius * sphere.radius;
 	float h = b * b - c;
 	if (h < 0.0) return false; // no intersection
-	h = sqrt(h);
+	h = sqrtf(h);
 	*out_pt = ray.pos + ray.dir * (-b - h);
 	return true;
 	//return vec2(-b - h, -b + h);
